Readjustments and fixes in the codebase
Bring back the ability to display results from multiple models in the
display_results.ipynb notebook. This now applies to both notebooks, the
ones in experiments/ and experiments/test-best-model. Also swapped the
columns and rows in the table for easier readability and overview.

Bring back the ner_today.ipynb notebook to show the "baseline" of
models. The results in the notebook show how well the modern written
language NerTagger model can handle 19th century Estonian.

Bring back the confusion matrices in the display_results.ipynb notebook.
The results now also show with which types of NE-s the model has trouble
with defining and instead defines it as another entity.